+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changelog
 ===
 
+## v0.2.0
+
+- Refactored cli stuff
+- Added subcommands `build` and `list`
+- Now also clone empty directories
+- Auto expand directory with placeholder default values if needed when building or listing
+
+## v0.1.2
+
+Small fixes
+
 ## v0.1.1
 
 Initial release, base features to get started.